refactor(hud): decompose GameplayHud into gameplay/ sub-components
Extract BoardStage, PlayerRosterPanel, PlayerHandPanel, SupplyDock,
InspectorDock, and notifications into dedicated files under
src/components/gameplay/. GameplayHud.tsx becomes a pure re-export
barrel so all existing consumers continue to work unchanged.

Uses updated Panel variants ("info", "private"), SectionHeader variant
prop, and Button "ghost" variant from the synced design system.
